base project : role middleware / post : post list , delete post , validation on create two user account with difference role and same email

+++ Modules/Post/routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Middleware\RoleMiddleware;
 use Illuminate\Support\Facades\Route;
 use Modules\Post\Http\Controllers\PostController;
 
@@ -10,6 +11,7 @@ Route::middleware(['auth:sanctum'])->prefix('v1')->group(function () {
 
 
 Route::prefix('posts')->controller(PostController::class)->group(function () {
-    Route::post('', 'store')->middleware('auth:sanctum');
-    Route::delete('/{post}', 'destroy');
+    Route::post('', 'store')->middleware(['auth:sanctum','role:editor']);
+    Route::delete('/{post}', 'destroy')->middleware(['auth:sanctum','role:admin,editor']);
+    Route::get('', 'index');
 });
